requirements.txt
django
djangorestframework
psycopg2-binary
gunicorn
django-filter
